Change strategy to provide `block_device` fixture on `win32`
The Hyper-V module does not seem to be installed by default on
GitHub-hosted runners, so we cannot create VHDX images on the fly using
`New-VHD`.

Because of that, we now commit gzip-compressed versions of the VHDX
images we need for testing purposes to version control and
mount/dismount these images using `Mount-DiskImage`/`Dismount-DiskImage`.