Fix calculation of block device size on Darwin systems
Seeking to the end of a device file does not seem to work on macOS like
it does on Linux. Instead, we should call `ioctl()` with
`DKIOCGETBLOCKCOUNT` to get the total amount of logical sectors on the
block device and multiply this value with the logical sector size of
the device.